@@ -33,7 +33,7 @@ typedef ProgressIndicatorBuilder = Widget Function(
 typedef LoadingErrorWidgetBuilder = Widget Function(
   BuildContext context,
   String url,
-  dynamic error,
+  Object error,
 );
 
 /// Image widget to show NetworkImage with caching functionality.
@@ -49,20 +49,20 @@ class CachedNetworkImage extends StatelessWidget {
   /// [BaseCacheManager] as the in memory [ImageCache] of the [ImageProvider].
   /// [url] is used by both the disk and memory cache. The scale is only used
   /// to clear the image from the [ImageCache].
-  static Future evictFromCache(
+  static Future<bool> evictFromCache(
     String url, {
     String? cacheKey,
     BaseCacheManager? cacheManager,
-    double scale = 1.0,
+    double scale = 1,
   }) async {
-    cacheManager = cacheManager ?? DefaultCacheManager();
-    await cacheManager.removeFile(cacheKey ?? url);
+    final effectiveCacheManager = cacheManager ?? DefaultCacheManager();
+    await effectiveCacheManager.removeFile(cacheKey ?? url);
     return CachedNetworkImageProvider(url, scale: scale).evict();
   }
 
   final CachedNetworkImageProvider _image;
 
-  /// Option to use cachemanager with other settings
+  /// Option to use cacheManager with other settings
   final BaseCacheManager? cacheManager;
 
   /// The target image that is displayed.
@@ -201,12 +201,15 @@ class CachedNetworkImage extends StatelessWidget {
   /// Will resize the image and store the resized image in the disk cache.
   final int? maxHeightDiskCache;
 
+  /// Listener to be called when images fails to load.
+  final ValueChanged<Object>? errorListener;
+
   /// CachedNetworkImage shows a network image using a caching mechanism. It also
   /// provides support for a placeholder, showing an error and fading into the
   /// loaded image. Next to that it supports most features of a default Image
   /// widget.
   CachedNetworkImage({
-    Key? key,
+    super.key,
     required this.imageUrl,
     this.httpHeaders,
     this.imageBuilder,
@@ -234,9 +237,10 @@ class CachedNetworkImage extends StatelessWidget {
     this.cacheKey,
     this.maxWidthDiskCache,
     this.maxHeightDiskCache,
+    this.errorListener,
     ImageRenderMethodForWeb imageRenderMethodForWeb =
         ImageRenderMethodForWeb.HtmlImage,
-  })  : _image = CachedNetworkImageProvider(
+  }) : _image = CachedNetworkImageProvider(
           imageUrl,
           headers: httpHeaders,
           cacheManager: cacheManager,
@@ -244,17 +248,17 @@ class CachedNetworkImage extends StatelessWidget {
           imageRenderMethodForWeb: imageRenderMethodForWeb,
           maxWidth: maxWidthDiskCache,
           maxHeight: maxHeightDiskCache,
-        ),
-        super(key: key);
+          errorListener: errorListener,
+        );
 
   @override
   Widget build(BuildContext context) {
     var octoPlaceholderBuilder =
         placeholder != null ? _octoPlaceholderBuilder : null;
-    var octoProgressIndicatorBuilder =
+    final octoProgressIndicatorBuilder =
         progressIndicatorBuilder != null ? _octoProgressIndicatorBuilder : null;
 
-    ///If there is no placeholer OctoImage does not fade, so always set an
+    ///If there is no placeholder OctoImage does not fade, so always set an
     ///(empty) placeholder as this always used to be the behaviour of
     ///CachedNetworkImage.
     if (octoPlaceholderBuilder == null &&
@@ -307,7 +311,10 @@ class CachedNetworkImage extends StatelessWidget {
       downloaded = progress.cumulativeBytesLoaded;
     }
     return progressIndicatorBuilder!(
-        context, imageUrl, DownloadProgress(imageUrl, totalSize, downloaded));
+      context,
+      imageUrl,
+      DownloadProgress(imageUrl, totalSize, downloaded),
+    );
   }
 
   Widget _octoErrorBuilder(
